# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Reference API
  version: 1.0.0
  summary: Shared HTTP contract implemented by both reference backends.
  description: |
    Design-first contract (see docs/architecture.md §4). Both
    `django/reference_api` and `fastapi/reference_api` implement it, and
    `tests/contract` verifies each running backend against it.

    Conventions:
    - Errors use RFC 9457 Problem Details (`application/problem+json`).
    - Every response carries `X-Request-ID`; send one to correlate client logs.
    - Timestamps are RFC 3339 in UTC. Money is integer minor units + ISO-4217.
    - Strings must not contain NUL (U+0000) or unpaired surrogates; both are
      rejected with 422 (`null_characters_not_allowed`,
      `surrogate_characters_not_allowed`) because PostgreSQL cannot store them.
    - Unknown body fields are rejected (422 `unknown_field`); unknown query
      parameters are ignored.
    - Text fields that must not be blank are trimmed with Python's definition of
      whitespace (`str.strip`), which is wider than JSON Schema's `\s` (it
      includes U+001C-U+001F and U+0085). Their `pattern` (anchor `non_blank`)
      lists exactly that set, so "schema-valid" and "accepted" agree.
  license:
    name: MIT
    identifier: MIT
# Every operation requires a bearer token unless it says otherwise
# (deny by default, like both backends).
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8410
    description: Django (local)
  - url: http://localhost:8420
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create an account.
      description: Returns the new user; sign in with `POST /api/v1/auth/token`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Account created.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/auth/token:
    post:
      tags: [auth]
      operationId: obtainToken
      summary: Exchange email + password for an access and a refresh token.
      description: Wrong email and wrong password give the same 401, so the endpoint does not reveal which emails exist.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenRequest" }
      responses:
        "200":
          description: Tokens issued.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/auth/token/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Rotate a refresh token.
      description: |
        Returns a new access token and a NEW refresh token. The submitted
        refresh token stops working (rotation), so a stolen token has a short
        useful life.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: Tokens rotated.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenPair" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke a refresh token.
      description: Access tokens are short-lived and expire on their own; logging out revokes the refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204":
          description: |
            Always 204 for a well-formed request: an unknown, expired or
            already-revoked token has nothing left to revoke, so logout is idempotent.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: The authenticated user.
      description: Clients call this after sign-in and on startup to restore the session.
      responses:
        "200":
          description: Current user.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `POST /api/v1/auth/token`. Expires after 15 minutes.
...
  headers:
    RequestId:
      description: Correlation ID. Echoes a valid incoming `X-Request-ID`, otherwise generated.
      schema: { type: string, pattern: "^[A-Za-z0-9._-]{1,128}$" }
  responses:
    Problem:
      description: Any error, as RFC 9457 Problem Details.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    ErrorCode:
      type: string
      description: Stable, machine-readable error code. Clients switch on this, never on `title`/`detail`.
      enum:
        - bad_request
        - validation_error
        - authentication_error
        - authorization_error
        - not_found
        - method_not_allowed
        - not_acceptable
        - conflict
        - idempotency_key_reused
        - unsupported_media_type
        - rate_limited
        - internal_error
        - external_service_error
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: [string, "null"]
          description: Dotted path to the field (`items.0.name`); null for errors not tied to one field.
        code: { type: string, description: "Validator code, e.g. `required`, `max_length`, `invalid`." }
        message: { type: string }
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { $ref: "#/components/schemas/ErrorCode" }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
        retry_after:
          type: integer
          minimum: 0
          description: Seconds until retrying is useful (only for `rate_limited`).
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, display_name, date_joined]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string, maxLength: 100 }
        date_joined: { type: string, format: date-time }
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 10, maxLength: 128, writeOnly: true }
        display_name: { type: string, minLength: 1, maxLength: 100, pattern: &non_blank '[^\t-\r\u001c-\u0020\u0085\u00a0\u1680\u2000-\u200a\u2028\u2029\u202f\u205f\u3000]' }
    TokenRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, minLength: 1, maxLength: 254, pattern: *non_blank }
        password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh]
      properties:
        refresh: { type: string, minLength: 1, maxLength: 1024, pattern: *non_blank }
    TokenPair:
      type: object
      additionalProperties: false
      required: [access, refresh]
      properties:
        access: { type: string, description: "Short-lived JWT for the Authorization header." }
        refresh: { type: string, description: "Long-lived, single-use token for POST /auth/token/refresh." }
...